🔧 chore(__main__.py): remove unused imports and functions to improve code cleanliness and maintainability 🔧 chore(main.py): update import statement to use get_number_of_workers from __main__ module 🔧 chore(util.py): remove unused imports and functions to improve code cleanliness and maintainability
